Create proper python package directory structure.
Rename the hyperspectral source code files to eliminate the
hyperspectral_ prefix and changed the name of terra_hyperspectral.py to
extractor.py.

Move the supporting scripts to a bin directory. Move the test program to the
tests directory. Move netCDF files to a data directory. Should the data files
bin the repository at all.